fix: bulk name setting for AxesTuple
